Break out pasting to provider to own function
This makes it easier to make custom pasting functions that will try a
spacific provider to begin with, it won't *only* try that provider
since the failover is the same as always which might be suboptimal, I
would like to make that an option so one can decide to only ever use
one provider and just fail if one wants to.

This is good if you have internal systems at offices for example, that
you have an internal paste provider for example github
enterprise. Then you don't want it to go try some other provider
afterwards.