Moved error-handling of failed requests out so providers must specify it
Moved error-lambda out, there's currently two predefined ones. One
that does allow failover and one that doesn't. This also kills the
newly introduced no-failover key. But it's for the better anyways.

This will make it much easier to actually do proper testing of
different cases in the code that simply wasn't possible before.

Also added a test of running the success lambda.