Added option to send returned URL to clpboard
If simpleclip is installed, users can send their returned URL to the
clipboard by setting webpaste/copy-to-clipboard to a non-nil value.

This change also adds webpaste/add-to-killring so that sending the URL
to the killring can be disabled, although to keep the same default
behavior the var's value is set to t by default.

The README is also updated to reflect these new options.