Begin autoloading of the webpaste group
This will make it much easier for the user to discover available
options. It seems like a bad thing to autoload defcustom's but not
defgroup's.

This conclusion is based on what's have been done in haskell-mode:

Here they remove their autoloads of customs because it breaks shit:
https://github.com/haskell/haskell-mode/commit/f0affb46b857c62d30de31a678d36e56831b775d

But, what they don't remove is the autoload of the group:
https://github.com/haskell/haskell-mode/blob/f0affb46b857c62d30de31a678d36e56831b775d/ghc-core.el#L35

So I'll assume that it's safe to autoload the group. This makes it
possible for users to run M-x customize-group webpaste without having
to either do a manual require or to use the mode before configuring.

This is a good improvement for discoverability for new users.